+++ asyncgit/src/sync/mod.rs
@@ -109,7 +109,10 @@ pub use utils::{
 	get_head, get_head_tuple, repo_dir, repo_open_error,
 	stage_add_all, stage_add_file, stage_addremoved, Head,
 };
-pub use worktree::{create_worktree, get_worktrees, WorktreeInfo};
+pub use worktree::{
+	create_worktree, get_worktrees, remove_worktree,
+	toggle_worktree_lock, WorktreeInfo,
+};
 
 pub use git2::ResetType;