+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * support x509 commit signing [[@kaden-l-nelson](https://github.com/kaden-l-nelson)] ([#2514](https://github.com/gitui-org/gitui/issues/2514))
 * worktrees popup: list worktrees and switch between them ([Shift+W])
 * worktrees popup: create a new worktree with a new branch ([c] from the worktrees list)
+* worktrees popup: remove a worktree ([Shift+D], refused if it is dirty or locked) and lock/unlock it ([l])
 
 ### Changed
 * use [tombi](https://github.com/tombi-toml/tombi) for all toml file formatting